Always show full description on checkout
